Update imagePadding calculation in encoders
Refactor the `imagePadding` property in the `LoadOptions` method of `Ean13Encoder.cs` and `Ean8Encoder.cs`. The calculation has been changed from a fixed value of `2 * this.Options.Scaling` to a more flexible approach using `this.Options.Margins * this.Options.Scaling`. This adjustment allows for better customization of padding based on user-defined margins.